Fixed GitHub Issue #17 - nova network failures
The Nova network create command now needs a label to create a network.
I've created a label parameter in the nova_network type that currently
defaults to novanetwork.

The nova network list command also now fails if no networks exist.
I've wrapped it in a rescue block which returns false is the command
fails which is clearly not the right response but I need this to work
for a demo.