'nova' user conflicts with package-created user
In Debian/Ubuntu, the 'nova' user is now created as a system user, which
makes sense.
The 'nova' user created in the 'nova' class is not created as a system
user and is created before the 'nova-common' package is installed and
prevents the installation of the package.
So, require the package to be installed before puppet checks the user
existence and create the 'nova' user as a system user.

Signed-off-by: François Charlier <francois.charlier@enovance.com>