Remove facter ec2 fact from example file
This is so that the nova example can be built in
ec2.

Currently, it re-reoutes calls to the metadata server
to its own internal server, thus breaking ec2 metadata on
ec2 instances running openstack.

This fact removes the facter source file so that puppet can still
run once a node is in this state.