set root_helper defaults
Starting is essex 2, a new utility has been created
that simplifies the management of sudoers config
for nova since nova runs as a non-root user but
often needs escalated privelages.

http://wiki.openstack.org/Packager/Rootwrap

The default setting for root_helper is sudo which
relies on a complex set of sudoers rules.

These rules can be greatly simplified by setting
root_helper to sudo nova-rootwrap.

This may be problematic b/c it depends on rather
or not this file has been packaged..

This config default has been added to the params
class b/c it depends on what OS packages are used
for installation.

Perhaps I can revisit this once the openstack
packaging matures.