Refactor nova services to use generic_service
Previously, there was a lot of copy/paste code for
handling the various nova services.

This commit creates the define nova::generic_service
which is used to capture common code for
configuring nova services.

It also updates the following classes to use that code:
  - nova::api
  - nova::cert
  - nova::compute
  - nova::objectstore
  - nova::network
  - nova::sceduler
  - nova::volume

It also updates spec tests for all of these classes