Implement exported resources for multi-node nova
This commit re-implements exported resources
to better support the deployment scenario where
nova compute nodes are split from the rest of
the deployment.

It removes the exported database connection.
The assumption is that all exported resources
should be explicitly performed by users implementing
their own modules.

It also removes references to cluster id used for collection.
This code is being removed b/c it is not fully implemented yet.
The basic design idea will later be reviewed and correctly implemented.

It ensures that conditional statements exist that determine rather
values have been supplied or if values for specific nova config should
be collected from the database.

The nova_config values that can possibly be collected are:
- rabbit_host
- glance_api_servers
- sql_connection