Refactor vnc code
This commit refactors the vnc implementation.

Refactors vnc related parameters to be coupled with
the services that need them. I am not entirely sure
about which service novncproxy_base is coupled with
so it is still defined in the nova class.

Adds the class nova::consoleauth.

Refactor vncproxy to install from source.