Move auth_strategy to nova class
This commit moves the auth_strategy config from
the api class to the nova class.

This is required b/c the compute node has to know
how to authorize for connections to the glance api
server.

Previously, the nova compute nodes were failing to
be able to communicate with glance on multi-node
installations.