Allow nova::rabbitmq to be disabled.
This commit adds the parameter enabled to the class nova::rabbitmq.
This parameter was added to allow this service to be deployed as
passive in an active-passive HA cluster.

This commit also adds test coverage to the nova::rabbitmq class.

It also adds the rabbitmq module as a dependency.y