nova manage network no longer refreshes db sync
Previously, the nova network creation caused the db migration
script to be notified. This code was added back in the Diablo days
and does not appear to be necessary.

It was removed b/c it made the db migration code really confusing
(resulting in two types that performed the migration)

This code removes the expectation from the test that this
exec will be refreshed.