change nova_config to use ini_file
This commit refactors the nova_config type to use
the ini_file provider. This will ensure that this type
is consistent with other types. It also completely removes
the parsed and parsed_config types.

The usage of inifile from nova_config is slightly difference
from other similar types. It does not require a section name,
and assumes that settings supplied without a section name
belong in the default section.

This is for backwards compatibility with the other types
and will likely change in the next major version.

this commit also updates nova_paste_int to support purging.