Remove quantum_connection_host parameter
quantum_connection_host parameter was not expected to work anyway
due to a typo in the logic in the class. quantum_connection_host
was removed from OpenStack with the release of Folsom.

It fixes a deprecation warning raised by Puppet:
Dynamic lookup of $quantum_host at nova/manifests/network/quantum.pp:27
is deprecated.

Additional changes:
* Remove commented parameters which were not used
* Add rspec tests for nova::network::quantum

Change-Id: I1d83511b0487de5ae694dae2b89bc265f6af27ca