Don't require flat_inferface.
The nova default for flat_interface is None but our puppet modules
currently require it to be set. This patch updates the flatdhcp
module so that it defaults to undef thus making it possible to use
nova::network w/ flatdhcp when no flat_interface(private_interface) is
defined.

Fixes LP Bug #1170569.

Change-Id: I62b91d9aef652c9d6b0a9e344832f6bef2bc3f27