Don't require flat_inferface.
The nova default for flat_interface is None but our puppet modules
currently require it to be set. This patch updates the network/flat
module so that it defaults to undef thus making it possible to use
nova::network w/ flat networking when no flat_interface(private_interface)
is defined.

Change-Id: I7167b76f95f3fe9a1ad208223def9597e56a4ecc