Manage the messagebus service on non-RHEL RHEL-like platforms
Previously, the test for enabling the "messagebus" service checked
for "$operatingsystem == RedHat", which only matches on RHEL;
$operatingsystem is "CentOS" on CentOS machines. I think we want
"$osfamily" here, which is "RedHat" on CentOS machines.

I also added a test for "$operatingsystem != Fedora" to preserve
the existing behavior on those machines, which as far as I can
tell do not require this service.

Change-Id: I020591006df87a9766cefbe36740940915924180