Adds Support for X-Forwarded-For HTTP Headers
Previously, the nova::api class did not support proxies and
load-balancers that use X-Forwarded-For Headers to identify the
originating IP address of requests.  This change implements the
use_forwarded_for parameter to enable/disable the ability for
Nova API to treat the X-Forwarded-For HTTP Header as the canonical
remote address.

Defaults to false for backwards compatibility and to disable
X-Forwarded-For Headers.

Change-Id: I3c7224bef1bfa4578ab2a298b05a24e46d618ce7