Fedora 18+ and EL 7+ feature more granular libvirt packaging
Fedora 18 or later and RHEL (or derivatives) 7 or later feature a
more granular libvirt packaging, thereby allowing to install only
just the necessary packages.

Most notably, this way no default libvirt network is set up. In the
past, the existence of an interface virbr0 with an IPv4 address of
192.168.122.1/24 often caused confusion with users when debugging
network issues.