Replace cinder::client class declaration with include
The nova::api class currently declares the cinder::client
class. Having this dependency set as a declaration leads
to issues when you try to combine the nova::api class
with other classes that also need to include the
cinder::client class (ie openstack::client).

This commit changes the declaration to an include to
ensure that this class can be combined with all other
classes without conflict.

Change-Id: I9c6604acfbde4c00629c9ab658f81b9444e85f7f