Removes unnecessary api-paste.ini configuration
Paremeter api_paste_config defaults to '/etc/nova/api-paste.ini',
so it is not needed to set it in nova.conf. Without having the parameter
in nova.conf we can use this module together with dist.conf. Also
auth parameters should be in nova.conf and not in api-paste.ini.

Change-Id: I85e4626076429d3e6955b7578320aac1f40d056a