Removes unnecessary api-paste.ini configuration
Paremeter api_paste_config defaults to '/etc/nova/api-paste.ini',
so it is not needed to set it in nova.conf. Without having
the parameter in nova.conf we can use this module together
with dist.conf. Also filter:authtoken parameters should be
in nova.conf and not in api-paste.ini. Removes filter:authtoken
configuration from api-paste.ini in case it is filled there.

Change-Id: I4798124b3c75c012653c9868c771146f23cf4cc7