Add the flat_interface parameter to nova::network.
By default the flat_interface is assumed to be private_interface,
but this might not be the case for certain scenarios. For instance,
one might want to have all VMs automatically bridged to a different
interface, such as a public interface, when your company or
institution's infrastructure is running on a flat public network.