Introduce nova::config + Update nova::config example description.
Add a new class nova::config which allows the creation and
management of arbitrary nova_config and nova_api_paste_ini resources.

This will add the ability for the end user to use Hiera to manage
those resources.

The original example in nova::config is syntax invalid.
Besides fixing that, this patch also gives an example in yaml format.

(cherry-picked from 40188356 and 133d5da1 in master)
Change-Id: Iad7a2b764884887a75f5f6db3e1660644e07c5c0
Closes-bug: #1290634