Ensure conductor_workers stay at previous value
During the refactoring of the conductor_workers parameter from nova::api
to nova::conductor, the conductor_workers default changed from
$::processorcount to the parameter being absent.

While this is consistent with the nova default behaviour, it changes the
value for people who currently have a different default set. The solution is
to do nothing, in case no value is given.

For reference see:
https://review.openstack.org/#/c/106656
https://review.openstack.org/#/c/108941
http://docs.openstack.org/trunk/config-reference/content/list-of-compute-config-options.html

Change-Id: I908fc4953522f0b1db8f43882a44407aba9fa9f9