crontab to move deleted instance rows to another database table
Create a crontab which moves deleted instances rows to another database table that you don't have to backup.
This is done to optimize database management at scale.
Depending on the amount of deleted instances that you get on a daily
basis you want to modify the $max_rows parameter.

This feature is implemented as in nova CLI and addressed by this
blueprint: https://blueprints.launchpad.net/nova/+spec/db-archiving

Change-Id: Ia03fa6862aefcbcc7d7949b08d00bacf679cb548
Signed-off-by: Emilien Macchi <emilien.macchi@enovance.com>