Add parameter to control use of rbd for the ephemeral storage
Currently the rbd class configures Nova so that it uses the
rbd driver for both the ephemeral and the persistent storage.
One might want to use it to attach Cinder volumes (persistent)
but not for the ephemeral storage instead, which be on local
disks or even backed by a different driver.

This change adds a parameter to control use of rbd driver for
the ephemeral storage, setting it to true by default to
preserve the old behavior.

Co-Authored-By: Giulio Fidente <gfidente@redhat.com>
Co-Authored-By: Emilien Macchi <emilien@redhat.com>

Change-Id: I4ae0fd605c5a57aa23bea83b06530a50844d24a0
Closes-bug: 1365554