Fix RedHat service provider conditional
Without this patch, params.pp conditionally uses 'RedHat' as a matching
condition to $::operatingsystem for two different cases, and will match
on the first case. This is a problem because it will prevent the proper
service provider from being assigned on RedHat operating systems with
major version < 7. This patch removes RedHat from the first case so that
RedHat operating systems will match on the second case and assign the
init service provider if the major version is < 7.

Change-Id: Id5c49cc11a4112c27621c72a7b2e6fcaf9da6aef