Allow auth_name and auth_name_v3 to be the same
Customize the resource name for keystone::resource::service_identity
resources, and explicitly specify the auth_name as a parameter to
the resource.

This allows for situations where auth_name and auth_name_v3 are
the same.  An example use case is where the user backend is Active
Directory, and only a single service account may be created to
run all Nova services.

Also add a check to ensure the nova and novav3 services are not
given the same name.  This situation can happen if the auth_name
and auth_name_v3 parameters are set to the same user, but a specifc
service_name or service_name_v3 parameter is not given and they
get their default values from auth_name and auth_name_v3.

Change-Id: I041aa76ef413ae4be4ffa29948292289544acb49
Closes-bug: 1407170