Correct docs on format for nova::policy data
Correct the format of the $policies parameter to nova::policy.
That parameter is passed directly to create_resources, so it needs
to be in a format acceptable to that.

Just updating docs here as anyone already using the policy classes
are likely not using the (incorrectly) documented format, anyway.

Change-Id: I774678f0b49ee05b6aee6a5565302377d8eae186
Closes-bug: 1409897
(cherry picked from commit 96b96f33939862543e62695b1dff14b80d893c33)