Add PCI Passthrough/SR-IOV support
The main issue with INI files, it doesn't support complex structures
which end up being declared as strings. This approach get that chain of
character and verify it's a valid JSON equivalent before assigning the
value.

Another driver is even though Hiera supports complex structure, we tend
to keep it simple when using Hiera interpolations.

https://wiki.openstack.org/wiki/SR-IOV-Passthrough-For-Networking#SR-IOV_Configuration

Change-Id: I27d43af13fcc80f41a16c31a0a6cfe112dc96acb