Remove Python Package Declaration
This change removes the package resource 'python' to prevent likely
duplicate resource declaration issues for module users, being that
python is most likely to be managed outside of any OpenStack profiles or
roles. Additionally, the declaration is superfluous, as the
'python-nova' package will pull in python as a dependency if required.

Change-Id: Ib4670bea09bb0435774aa0835abb3fbb1ccc77e2