Allow libvirt secret key setting from param
Currently the libvirt secret key is demanded to the ceph cluster
via the $(ceph auth get-key ...) command which requires the ceph
cluste to be already up and also assumes the computes are provisioned
with the client.admin keyring.

With this change we add a libvirt_rbd_secret_key parameter which,
if passed, is used instead so that computes can be configured
without the distribution of additional keys.

Change-Id: I70da06159c0d3c6fa204b5f7a468909ffab4d633
Closes-Bug: #1439949
(cherry picked from commit a8ba5a41cdbcbc5b4123cc97e0e46f4c7702b8fa)