Add vagrant nodeset and make it the default
Added an Ubuntu nodeset stolen from the puppetlabs apache module.
The box_url is necessary even though the box is namespaced, vagrant
seems to have changed recently and just having the vagrant cloud
namespace in the box name is not enough to import the box.

We're making this nodeset the default so that a developer can have
it "just work" on their laptop. CI will be configured to use a
nodepool nodeset. Following the example of the puppetlabs modules,
the default nodeset is a copy of the new ubuntu nodeset so that
setting BEAKER_set=ubuntu-server-1404-x64 still works.

Fixed an error in the README (correct environment variable is
BEAKER_set not BEAKER_setfile) and added more details.

Change-Id: Ib3e50ffbaf613dd2e6582904efeff1d7cce77bb5