Run db-sync if nova packages are upgraded
This changes the db-sync logic to run if the api or common package is
upgraded.  Previously it would solely ensure that the api package was
installed before running db sync.  On ubuntu, if the common package is
upgraded then the api package will also be upgraded in the same
transaction, which means that no notify will be generated.

Lastly, this fixes the package collector to use the params names for the
packages, instead of hard coding nova-api

Change-Id: I4fcea354c5d663644592093ec617ee5b1c6f33e3