Better handling of package dependencies in nova generic_service
Before defining the package for the generic_service, check if it is
already defined, either by $noval_title or $package_name naming.
Create dependency relationship(s) with the service resource based
on how the package is defined in the catalog.

Otherwise, it's possible to end up with a duplicate package definition,
or, more likely, the dependency from the service definition to break.
This occurs most specifically with the spicehtml5proxy class on
EL-based platforms.

Change-Id: I215a95cac1d4b97837b67b417ff5fd7e6e9480d6
Closes-bug: 1465433