Creation of nova::db::sync
In order to standardize the way dbsync are run across our modules,
we create a new class nova::db::sync.
This class will be included if sync_db is enabled.

By making this transition the nova::db::sync class
can be returned by the ENC.

A use case would be in an highly available environment, with 3 galera
nodes, include nova::api on every node with sync_db set to false
and have the ENC return nova::db::sync just for one node.

Change-Id: I795178f83fe5b997a2574fe17263643688efc48c