Create networks only if nova-network is enabled
The problem is that if we use nova-network and multi host configuration, [1,2]
nova-network service can be not run yet. This is a problem because nova_network
provider uses nova client and then as nova-api just does an RPC call to
nova-network, which actually creates networks [3]. This impacts multinode
configurations with nova-network services running only on compute node.

[1] http://docs.openstack.org/openstack-ops/content/network_design.html
[2] http://docs.openstack.org/openstack-ops/content/example_architecture.html
[3] https://github.com/openstack/nova/blob/master/nova/network/api.py#L103

Closes-bug: #1501767

Change-Id: I44d78e946c0e01e883c6d54bb54054200aae4aee