Move cinder_catalog_info to init
In a previous commit, we introduced cinder_catalog_info parameter but it
was a mistake to configure it on nova api class, because this parameter
is also used by nova compute.

cinder_catalog_info is a parameter used by nova/volume/cinder.py in
cinderclient method, consummed by nova compute manager code.

This patch is non-backward compatible but it does not matter because it
was not in the latest stable release and has not been backported.

Change-Id: Id9e579ca31364d5207d0c1b892d0f7aa7f20f7a8