Allow spice protocol to be deployed
It's not possible to deploy spice on Nova Compute, as the default
configuration (even without VNC) pollutes the configuration files with
the two VNC settings:

 - DEFAULT/vncserver_proxyclient_address
 - DEFAULT/vnc_keymap

Thus, a conflict with deployment of non-VNC consoles is removed.

This change configures the two parameters if $vnc_enabled is set,
and otherwise removes them.

Also update the specs in nova_compute_spec.rb to handle the new truths
regarding a vnc that is disabled or enabled.

Change-Id: Ifcc619a9f9c0d31bc2e94ca0083433a94cb66bdd
Closes-Bug: #1518580
Signed-off-by: Martin Millnert <martin@millnert.se>