rbd: make sure libvirt is run before get-or-set virsh secret
If we try to get-or-set virsh secret before starting libvirt, we got
this:
Failed to connect socket to '/var/run/libvirt/libvirt-sock'

Also, drop the requirement on nova::config::end Anchor for libvirt
service, that already needs to run before nova-compute service.
It's created a dependency cycle:
Anchor[nova::config::end] => Service[libvirt] => Exec[get-or-set virsh
secret] => Exec[set-secret-value virsh] => Anchor[nova::config::end]

Also, stop requiring ceph-client-package Package resource in the
'get-or-set virsh secret' Exec, since the resource is not required in
this class.

This patch is an attempt to prevent this situation and allow to deploy
nova-compute with RBD in a single Puppet run.

Change-Id: I69a7d40e7d1847be06a843986ace4f0602272fe1