Add other virt services
This change adds the management of virtlockd and virtlogd services
to the nova::compute::libvirt. These services are actually required
for nova to be able to communicate correctly with libvirt.  Virtlockd
exists for RedHat and Ubuntu but is not currently available under
Debian. Virtlogd is Ubuntu specific service that is usually started
automatically on install of the libvirt package. However if you
disable that or some how prevent it from running, then it may
interfere with nova.  We should explicitly manage them since they
are required.

Change-Id: I52c6609e547fa50dccb4b31b9d93748b2b02dee0