Revert "Fix default keystone auth_plugin for nova/neutron"
This reverts commit 271a2587a38c6a396213c95aaa8a690fcc283091.

This patch seems to have caused the TripleO upgrades job to fail
in that running Nova instances stopped working after an upgrade.

Could be there is a missing relationship or something in Puppet or
perhaps something else keystone related but updating to this
as a default seems it could break things.

Related bug: #1580076

Change-Id: I783fba6cbd3ba292d6ed64be2dd0772e20a0896d