Add transport_url and rpc_response_timeout for oslo.messaging
This commit adds the transport_url parameters for oslo.messaging. The
url is of the form:

  transport://user:pass@host1:port[,hostN:portN]/virtual_host

Where the transport scheme specifies the rpc or notification backend
as one of rabbit, amqp, zmq, etc. Oslo.messaging is deprecating the
host, port, and auth configuration options [1]. All drivers will get
these options via the transport_url.

The rpc_response_timeout was also included because it is used
explicitly in oslo.messaging, and it's being set in some deployment
tools. So by using the oslo::messaging resource in nova, it will
break those deployment tools. So a commit to take this into use is
needed on that side.

This patch:
* use oslo::messaging::default resource
* use oslo::messaging::notifications resource
* add parameters for transport_url(s)
* add parameter for rpc_response_timeout
* update spec tests
* add feature release note

[1] https://review.openstack.org/#/c/317285/

Change-Id: I961e7b4f061efef3ee988e698b41a7241afb4053