Enable compute node micro-services
nova::compute::libvirt was a class that deployed both libvirt (+
external services related to it) and nova-compute bits (mainly
configuration).

In the micro-services use-case, we want individual services, that can
run alone in containers. To allow it, we need to split the nova compute
configuration when libvirt is configured and the libvirt deployment.

This patch aims to create nova::compute::libvirt::services that contain
the bits from nova::compute::libvirt that used to deploy Libvirt and
some other packages / services related to it.

We keep backward compatibility by declaring the new class in
nova::compute::libvirt but allow to disable it and select what we
actually need thanks to parameters (we support hiera & non-hiera, see
code for comments that document it).

This is a first iteration of micro services for Compute nodes, soon
we'll also work on nova::migration class to separate nova-compute &
libvirt bits again.

Change-Id: Ib0d3111560af5af451e522c6dc3b3918d0463e7d